Move JSON response handling into trait, ready for reuse
