+++ src/Api/JsonClient.php
@@ -13,6 +13,7 @@ namespace Mxm\Api;
 class JsonClient implements \Psr\Log\LoggerAwareInterface
 {
     use ConnectionTrait;
+    use JsonTrait;
 
     /**
      * @var string
@@ -124,20 +125,7 @@ class JsonClient implements \Psr\Log\LoggerAwareInterface
         $parts   = preg_split('|(?:\r?\n){2}|m', $response, 2);
         $content = $parts[1];
 
-        if ((int)$code != 200) {
-            try {
-                $message = $this->decodeJson($content);
-                if ($message instanceof \stdClass && isset($message->msg)) {
-                    $content = $message->msg;
-                }
-            } catch (\UnexpectedValueException $e) {
-                // Void
-                // Failed to decode, leave content as the raw response
-            }
-            throw new \RuntimeException($content, $code);
-        }
-
-        return $content;
+        return $this->processJsonResponse($content, $code);
     }
 
     /**
@@ -158,44 +146,6 @@ class JsonClient implements \Psr\Log\LoggerAwareInterface
     public function getLastResponse()
     {
         return $this->lastResponse;
-    }
-
-    /**
-     * Decode JSON
-     *
-     * @param string $json
-     * @return mixed
-     * @throws \UnexpectedValueException
-     */
-    protected function decodeJson($json)
-    {
-        $result = json_decode($json, false);
-
-        if ($result === null) {
-            // Error checking
-            switch (json_last_error()) {
-                case JSON_ERROR_DEPTH:
-                    $error = 'Maximum stack depth exceeded';
-                    break;
-                case JSON_ERROR_CTRL_CHAR:
-                    $error = 'Unexpected control character found';
-                    break;
-                case JSON_ERROR_SYNTAX:
-                    $error = 'Syntax error, malformed JSON';
-                    break;
-                case JSON_ERROR_NONE:
-                default:
-                    // Value is really null
-                    $error = '';
-                    break;
-            }
-
-            if (!empty($error)) {
-                throw new \UnexpectedValueException("Problem decoding json ($json), {$error}");
-            }
-        }
-
-        return $result;
     }
 
     /**
